change directory of MapConverter
